[distort] revert the previous commit and represent the image size as an integer

+++ src/distort-omp/main.cpp
@@ -36,7 +36,7 @@ int main(int argc, char **argv)
   prop.xscale = (prop.width - prop.xshift - xshift_2) / prop.width;
   prop.yscale = (prop.height - prop.yshift - yshift_2) / prop.height;
 
-  const size_t imageSize = (size_t)height * width;
+  const int imageSize = height * width;
   const size_t imageSize_bytes = imageSize * sizeof(uchar3);
 
   uchar3* h_src = (uchar3*) malloc (imageSize_bytes);
@@ -45,7 +45,7 @@ int main(int argc, char **argv)
   struct Properties *h_prop = &prop;
 
   srand(123);
-  for (size_t i = 0; i < imageSize; i++) {
+  for (int i = 0; i < imageSize; i++) {
     h_src[i] = {rand() % 256, rand() % 256, rand() % 256};
   }
 
@@ -66,7 +66,7 @@ int main(int argc, char **argv)
   // verify
   int ex = 0, ey = 0, ez = 0;
   reference(h_src, r_dst, &prop);
-  for (size_t i = 0; i < imageSize; i++) {
+  for (int i = 0; i < imageSize; i++) {
     ex = max(abs(h_dst[i].x - r_dst[i].x), ex);
     ey = max(abs(h_dst[i].y - r_dst[i].y), ey);
     ez = max(abs(h_dst[i].z - r_dst[i].z), ez);
